+++ Examples/Main.hs
@@ -3,9 +3,11 @@ module Main where
 import Examples.AssertCheckTest
 import Test.MuCheck.TestAdapter.AssertCheck
 
-main = do
-    assertCheckResult sortEmpty
-    assertCheckResult sortSorted
-    assertCheckResult sortRev
-    assertCheckResult sortSame
-    assertCheckResult sortNeg
+main :: IO ()
+main = mapM_ assertCheckResult
+    [ test_sortEmpty
+    , test_sortSorted
+    , prop_sortIsIdempotent
+    , sortSame
+    , test_sortNeg
+    ]